+++ src/include/sll/gc.h
@@ -38,7 +38,7 @@ __SLL_EXTERNAL void sll_remove_debug_data(sll_runtime_object_t* o);
 
 
 
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_return_t sll_verify_runtime_object_stack_cleanup(void);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_verify_runtime_object_stack_cleanup(void);