+++ src/sll/platform/windows.c
@@ -340,7 +340,7 @@ __SLL_EXTERNAL void sll_platform_socket_init(void){
 
 
 
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_return_t sll_platform_socket_execute(const sll_string_t* h,unsigned int p,const sll_string_t* in,sll_string_t* o){
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_platform_socket_execute(const sll_string_t* h,unsigned int p,const sll_string_t* in,sll_string_t* o){
 	sll_string_create(0,o);
 	struct addrinfo ah;
 	sll_zero_memory(&ah,sizeof(struct addrinfo));
@@ -356,14 +356,14 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_return_t sll_platform_socket_execute(const
 		p/=10;
 	} while (p);
 	if (getaddrinfo(h->v,ps+i+1,&ah,&r)){
-		return SLL_RETURN_ERROR;
+		return 0;
 	}
 	SOCKET s=INVALID_SOCKET;
 	for (struct addrinfo* k=r;k;k=k->ai_next){
 		s=socket(k->ai_family,k->ai_socktype,k->ai_protocol);
 		if (s==INVALID_SOCKET){
 			freeaddrinfo(r);
-			return SLL_RETURN_ERROR;
+			return 0;
 		}
 		if (connect(s,k->ai_addr,(int)k->ai_addrlen)!=SOCKET_ERROR){
 			break;
@@ -372,11 +372,11 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_return_t sll_platform_socket_execute(const
 	}
 	freeaddrinfo(r);
 	if (s==INVALID_SOCKET){
-		return SLL_RETURN_ERROR;
+		return 0;
 	}
 	if (send(s,in->v,in->l,0)==SOCKET_ERROR){
 		closesocket(s);
-		return SLL_RETURN_ERROR;
+		return 0;
 	}
 	sll_char_t bf[4096];
 	int l=recv(s,bf,4096,0);
@@ -384,7 +384,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_return_t sll_platform_socket_execute(const
 	while (l){
 		if (l<0){
 			closesocket(s);
-			return SLL_RETURN_ERROR;
+			return 0;
 		}
 		sll_string_increase(o,l);
 		sll_copy_data(bf,l,o->v+o->l);
@@ -394,7 +394,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_return_t sll_platform_socket_execute(const
 	closesocket(s);
 	o->v[o->l]=0;
 	sll_string_calculate_checksum(o);
-	return SLL_RETURN_NO_ERROR;
+	return 1;
 }